upgrade selecting
Clicking will now have a small radius around it, so that small things (like path points) are easier to click.

I also added the ability to right click a path point, and delete it (as well as using the delete key). Currently it does not automatically decrease the other path point numbers, but that may come later.

I also made sure that using the arrow keys does not move the selection in the object selector (by just shifting focus to the level canvas after clicking).